BMB critical value for the maximum of Chatterjee's rank correlations

This function returns the bootstrap critical value of the statistic $$T^B=\max_{1\le j\le p}\frac{1}{\sqrt{mq}}\sum_{k=1}^{m} \hat{A}_{j,k} \cdot \varepsilon_{k},$$ where the \(\varepsilon_{k}\) are independent standard normal random variables, based on \(B\) block multiplier bootstrap samples.

Usage

BMB.cv(dat, q, B, alpha, type = "bmb1", seed = NULL)

Arguments

dat

List with two elements: vector X and matrix Y. The \(j\)-th column in \(Y\) corresponds to \(Y_j\) in the individual hypothesis \(Y_j \perp X\).

q

Numeric. The block size.

B

Numeric. The number of bootstrap samples.

alpha

Numeric. The significance level.

type

Character. This argument specifies whether and how the test statistic and the bootstrap statistic are studentized. Options are "bmb" (no studentization), "bmb1" (default option), and "bmb2" (alternative studentization). The types are formally described in Olivares, Olma, and Wilhelm (2025).

seed

Numeric. The seed for the random number generator. If NULL, the seed is not set. If a positive integer, it sets the seed for reproducibility.

Value

Returns a list with the critical value and the type of test statistic.

References

Olivares, M., Olma, T., and Wilhelm, D. (2025). A Powerful Bootstrap Test of Independence in High Dimensions. Preprint, arXiv:2503.21715.

Author

Maurcio Olivares

Tomasz Olma

Daniel Wilhelm

Examples

if (FALSE) {
n <- 100
p <- 10
dat <- list()
dat$X <- rnorm(n)
dat$Y <- MASS::mvrnorm(n = n, mu = rep(0,  p), Sigma=diag(rep(1, p)))
BMB.cv(dat, q=2, B=100, alpha=0.05, seed = 5) # default type is "bmb1"

}